+++ exonum-light-client/src/test/java/com/exonum/client/SystemApiHelperTest.java
@@ -23,6 +23,7 @@ import static org.junit.jupiter.params.provider.Arguments.arguments;
 
 import com.exonum.client.response.ConsensusStatus;
 import com.exonum.client.response.HealthCheckInfo;
+import com.exonum.client.response.SystemStatistics;
 import com.google.common.collect.ImmutableList;
 import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -39,18 +40,28 @@ class SystemApiHelperTest {
   }
 
   @ParameterizedTest
-  @MethodSource("memoryPoolSource")
-  void parseMemoryPoolJson(String json, int expected) {
-    int actual = SystemApiHelper.parseMemoryPoolJson(json);
+  @MethodSource("statsSource")
+  void parseStatsJson(String json, SystemStatistics expected) {
+    SystemStatistics actual = SystemApiHelper.parseStatsJson(json);
     assertThat(actual, is(expected));
   }
 
-  private static List<Arguments> memoryPoolSource() {
+  private static List<Arguments> statsSource() {
     return ImmutableList.of(
-        arguments("{\"size\": 0}", 0),
-        arguments("{\"size\": 2}", 2),
-        arguments("{\"size\": " + Integer.MAX_VALUE + "}", Integer.MAX_VALUE)
+        arguments(formatStats(0, 0), new SystemStatistics(0, 0)),
+        arguments(formatStats(1, 2), new SystemStatistics(1, 2)),
+        // Check deserializer accepts longs-as-numbers (which is non-standard, but used by Exonum)
+        arguments(formatStats(Integer.MAX_VALUE, Long.MAX_VALUE),
+            new SystemStatistics(Integer.MAX_VALUE, Long.MAX_VALUE)),
+        // and longs-as-strings (a common approach to represent longs)
+        arguments("{\"tx_pool_size\": " + 0 + ", \"tx_count\": \"" + Long.MAX_VALUE
+                + "\"}",
+            new SystemStatistics(0, Long.MAX_VALUE))
     );
+  }
+
+  private static String formatStats(int txPoolSize, long txCount) {
+    return String.format("{\"tx_pool_size\": %d, \"tx_count\": %d}", txPoolSize, txCount);
   }
 
   private static List<Arguments> healthCheckSource() {